Implement a configurable reply plugin
Rather than writing custom C code every time you want to implement some
replies, might as well shove the replies into configuration files. This
implementation allows you to specify one or several responses (randomly
selected), whether the bot should be addressed or not, and the regex
trigger.

To ensure that the configurations are flexible, they should be able to
specify responses which include the user who sent the message, as well
as other contextual strings. The response also should be able to include
captured strings from the regex. To do this, we implement a string
formatter with syntax similar to Python's f-strings.

This can supersede the functionality of magic8, lod, and sadness
plugins, so they are removed. This implements #17 and inadvertently
closes #13 by deleting that code. Sorry @angelolicastro.