irc: fix missing channel members for some channels
We expected a TOPIC reply for joining a channel, but sometimes that
didn't happen. Stop expecting the TOPIC reply. Just expect a NAMREPLY.
If we get a TOPIC, update it, but don't create any state which is
waiting for this. This means we can get rid of the "join request"
structure.

This revealed a database query error for clearing memberships: don't
use a join in a DELETE, move it to a subquery.