signal: Properly handle mentions and usernames
Signal represents a mention with a unicode code point 65532, object
replacement. Alongside it there is a JSON array of the actual
replacement values. To simplify things, we simply translate that into a
format easy to use: @(uuid:xxxxxxx). This internal representation will
mostly pass through plugins unscathed, and can be translated by the send
function back into a signal mention.

To sweeten the deal, we can also encode the user and group in a similar
format. The result, is that plugins which insert user "nicknames" into
their messages will end up inserting a "mention placeholder" and that
will get automatically translated into the correct API mumbo jumbo.

Unfortunately, this does add significant complexity. We need to do a
fair bit of escaping and unescaping to make it all work. I'm sure there
are subtle bugs, not to mention casual memory leaks here.