Enable proper thread cleanup
Exiting without ensuring all sc-lwt threads have terminated is a sure
way to have a memory leak (which admittedly doesn't matter, since you're
exiting). It's also a way to potentially leak other resources without an
opportunity to cleanup. This is not ideal.

In v0.6.0 of sc-lwt a few primitives are introduced to allow proper
cleanup of threads:
1. sc_lwt_send_shutdown_signal() can wake every task, and responsible
   threads will check sc_lwt_shutting_down(), exiting gracefuly when
   it returns true.
2. sc_lwt_join() will allow a task to wait until it is the only
   remaining thread.
3. sc_lwt_task_count() allows a task to check how many tasks exist in
   the scheduler, ensuring that they are the last one there (or forcing
   an early shutdown otherwise).

Use these tools in src/cbot.c so that when backends finish running, we
gracefully shutdown any other thread (with 5-second timeout), and then
shut down the bot application itself.