Fetch TAG from GITHUB_REF
GITHUB_REF comes populated as "refs/heads/master" for builds triggered
by normal commits. These should skip all of the steps which concern
creating releases. For pushed tags, it comes populated as
"refs/tags/vX.Y.Z". If you tag a release, you'll get one of each build,
so it's important that the build knows to only try to release when it's
the build for the real tag, not just the commit.