signal: allow jmsg_wait() and friends to be called from other threads
The jmsg_wait() and sig_result() functions all need to wait for a
confirmation from the signald socket, and to get that, they need to read
from the socket. However, only one thread can really be polling to
read from that socket.

To handle this, jmsg_wait_field() will only read from the socket if it
is on the main thread. Otherwise, it will enqueue a descriptor of a
signal_queued_item, which indicates that it would like to receive the
next JSON message which matches a certain key=value criteria.

In the main loop of the Signal backend, check these criteria and hand
off any matching message prior to handling that message.

Signed-off-by: Stephen Brennan <stephen@brennan.io>