birthday: Use cbot callbacks rather than thread
The birthday plugin was a pioneer in shitty uses of sleep / scheduling.
It "works" but wow is it dumb. The callback API is my new take on it
after thinking more. The API supports the use case for the trivia
plugin, which has more complex requirements (every week, two messages,
with other interactions complicating things). So it can trivially
support the birthday plugin.

So ditch the thread with the sleep loop. Simply schedule a callback.
This is so much easier to read!

Signed-off-by: Stephen Brennan <stephen@brennan.io>