trivia: Add support for non-email messaging
The code so far expects that you delegate to a sendmail command to send
the final message. However, I've had some trouble with getting that
email to go through on the other side.

SMS would work great, but I'm not going to do a whole Twilio thing for
this. I really just want to have my own phone send the SMS. So I rigged
up an automation in Tasker which connects to a ntfy.sh topic and sends
the SMS on my phone automatically. This way the bot does send the
message, but it comes from my phone and Grace can reply to me.

Signed-off-by: Stephen Brennan <stephen@brennan.io>