Rely on GIT_REF from sourcehut builds
GIT_REF is always set to the ref that triggered the build, in the case
of a release tag, we can detect that directly.

Polish up the POSIX-foo a bit here to create some /bin/sh compliant
string matching. Use the case statement to match against the tag or exit
otherwise. Use the parameter expansion ${VAR#prefix} to strip out
prefixes. This is used liberally to get rid of the "v" prefix of tags.

Signed-off-by: Stephen Brennan <stephen@brennan.io>