signal: add support for signal-cli bridge
The signald project is great, but has recently dropped behind on keeping
up with the Signal protocol, resulting in cbot missing messages. An
alternative is available, which is quite similar: signal-cli. Signal-cli
includes a jsonRpc command which presents an API that we can use just
like signald.

Introduce a signal "bridge" setting with suitable callbacks to implement
the API. Split out the signald-specific code into its own file, and
introduce a new signal-cli bridge. This code has been briefly tested: it
can send and receive messages in groups and DMs, and it can properly
handle reactions. Some bugs certainly remain to be found.

Signed-off-by: Stephen Brennan <stephen@brennan.io>