signal: fixes for is_authorized and addressed
Signal seems to now primarily use UUIDs, and the phone numbers are not
visible in profiles. Further, signal-cli doesn't give us any way to view
our own profile information (signald does). For both of these
limitations, we now need the config file to specify the UUID of the
bot's user, and we adjust the config to specify the UUID of the
authorized user as well.

With this change, we also factor out the bot alias setup from signald
code into the common signal code. That allows signal-cli to get the
proper detection of all forms of addressing the bot:

   cbot: ...   (plain text)
   @cbot: ...  (mentioned)
   @cbot: ...  (@ sign, plain text)
   @@cbot: ... (escaped @ sign, plain text)

Signed-off-by: Stephen Brennan <stephen@brennan.io>