signal: code cleanup and standardization
Over time the signal code has grown and gotten over-complicated,
especially with respect to the JSON parsing (jmsg) and extra code
related to functionality that was never used. This commit goes over all
of that and drops whatever we can, removes header declarations and
converts to static where possible, and migrates to the "je_" functions
of JSON parsing which I introduced recently.

Signed-off-by: Stephen Brennan <stephen@brennan.io>