sqlknow: use non-greedy matching
See, for instance:

> cbot know that it is what it is buddy
[stdin]cbot: ok, it is what it is buddy
> cbot what is it
[stdin]cbot: Sorry, I don't know anything about it
> cbot what is it is what it
[stdin]cbot: it is what it is buddy

Clearly, the intent is for the word "it" (whatevery comes before the
first "is") to be used as the knowledge key. Since the capture is
greedy, we instead get everything up until the last "it".

Switching to non-greedy is better; I can't think of a time where we'd
want greedy. With this change:

> cbot know that it is what it is buddy
[stdin]cbot: ok, it is what it is buddy
> cbot what is it
[stdin]cbot: it is what it is buddy

Signed-off-by: Stephen Brennan <stephen@brennan.io>