cli: put history file in working directory
For a general purpose tool, this would be quite rude: littering the
filesystem with .cbot_history files is not a kind thing. But the cbot
CLI is really only used in development.

Keeping the history file in the working directory ensures that the
Alpine container can access it, so I can take advantage of command
history from my test runs outside the container.

Signed-off-by: Stephen Brennan <stephen@brennan.io>