cbot: search the same directory as executable for plugins
With meson, the build directory contains the executable and all the
plugins. Let's default the plugin directory to the parent directory
containing the executable. By doing this, we can eliminate the
"plugin_dir" config from development configs. This means we can use the
same "cli.cfg" within the Alpine container as outside.

Signed-off-by: Stephen Brennan <stephen@brennan.io>