Add alpine build & test script
Developing on Arch with glibc, and then deploying an Alpine container
with musl... it's just stupid. I probably don't need a container to
begin with, but that's a problem for future me to deal with.

In the meantime, I'll be well served to have a local development
environment that's close to the production one that uses Alpine. Create
a Dockerfile for that, and an script to build & run within that
container.

Signed-off-by: Stephen Brennan <stephen@brennan.io>