sports_schedule: go back to using %I in strftime()
Turns out that %l is not available in Musl's strftime(). I would really
like it if there was an easy document on what *is* included! Seems that
in the manual page I'm reading, half of the format codes are
unavaialable.

We can skip over the leading zero instead of space, getting the same
behavior as desired for single-digit hours.

Also, let's start checking the return value of strftime() -- a novel
idea, I know. Now that the format string is fixed, it probably won't
fail anymore, but better safe than sorry.

Signed-off-by: Stephen Brennan <stephen@brennan.io>