sports_schedule: get rid of %z at end of strptime
The MLB dates include a "Z" at the end indicating "Zulu" aka UTC. I
assumed this was a time zone specifier which could be parsed by %z or
%Z. And hey, glibc did it, so I assumed my assumption was correct!

But unsurprisingly, I was wrong. Or at least, musl disagreed. So now, I
just assert the Z is at the end of the timestamp, which is probably best
anyway, because the subsequent code does assume that the timestamp is
UTC.

Signed-off-by: Stephen Brennan <stephen@brennan.io>