sports_schedule: switch to chasecentercalendar.com
The existing plugin only covered MLB/NBA games. It was nice that it was
"general purpose" so you could use it for other teams, but we did not
need that functionality. Limiting to MLB & NBA meant that events like
concerts, as well as the Valkyries' games, were not included. This
definitely resulted in some folks being caught off guard by event
traffic.

The chasecentercalendar.com project (albertyw/chase-center-calendar)
combines multiple sources of data and exposes them as iCal. Let's use
that to improve our output. The name "sports_schedule" is now a bit
outdated, which we'll address next.

Signed-off-by: Stephen Brennan <stephen@brennan.io>