tests: introduce plugin testing
This adds a cbot backend which can be used to help test plugins. It
allows creating a bot with the test backend, registering the plugin
under test, sending messages to the bot, and then fetching response
messages from the plugin.

Many things are yet missing: a mocked backend for HTTP requests, the
ability to run LWTs and ensure they complete, the ability to mock the
C time APIs and play with time, the ability to provide configuration to
a plugin, etc.

The end goal is to be able to test a complex plugin like the trivia
plugin. But for now, let's settle for testing a very simple plugin: the
name plugin, which just responds to questions about cbot's name.

Signed-off-by: Stephen Brennan <stephen@brennan.io>